requirements.txt
Package            Version
------------------ ---------
autopep8           1.5.7
beautifulsoup4     4.10.0
certifi            2021.5.30
charset-normalizer 2.0.5
comtypes           1.1.10
docopt             0.6.2
idna               3.2
Js2Py              0.71
MouseInfo          0.1.3
packaging          21.0
Pillow             8.3.2
pip                21.2.4
pipwin             0.5.1
PyAudio            0.2.11
PyAutoGUI          0.9.53
pycodestyle        2.7.0
PyGetWindow        0.0.9
pyjokes            0.6.0
pyjsparser         2.7.1
PyMsgBox           1.0.9
pyparsing          2.4.7
pyperclip          1.8.2
pypiwin32          223
PyPrind            2.11.3
PyRect             0.1.4
PyScreeze          0.1.28
pySmartDL          1.3.4
python-dotenv      0.19.0
pyttsx3            2.90
pytweening         1.0.4
pywhatkit          5.1
pywin32            301
requests           2.26.0
setuptools         51.1.2
six                1.16.0
soupsieve          2.2.1
SpeechRecognition  3.8.1
toml               0.10.2
tzdata             2021.1
tzlocal            3.0
urllib3            1.26.6
wheel              0.36.2
wikipedia          1.4.0
Wikipedia-API      0.5.4